chore(refactoring): remove useless computed property
Before this commit the element was using a computed property to (re)load
resources when the language is changed. I replaced this computed
property by an observer. The function to compare countries by names was
also replace by the `String.prototype.localeCompare` method (that does
exaclty the same thing but considering the locale).